% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual-plotFunctions.R
\name{calculate_cumulative_dlt}
\alias{calculate_cumulative_dlt}
\title{Calculate cumulative DLT proportion}
\usage{
calculate_cumulative_dlt(data, drugA, drugB, cohort_limit, expand_grid = FALSE)
}
\arguments{
\item{data}{Data frame containing patient outcomes}

\item{drugA}{Drug A object}

\item{drugB}{Drug B object}

\item{cohort_limit}{Maximum cohort to include}

\item{expand_grid}{Whether to include all possible combinations}
}
\value{
Data frame with cumulative DLT proportions
}
\description{
Calculates cumulative dose-limiting toxicity proportion for each dose combination
up to a specified cohort.
}
