% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies-admissible-Adjacent.R
\name{calculateAdjacentDoses}
\alias{calculateAdjacentDoses}
\title{Calculate Adjacent Doses}
\usage{
calculateAdjacentDoses(vec, nrows, ncols)
}
\arguments{
\item{vec}{Binary vector representing the current dose configuration}

\item{nrows}{Number of rows in the dose matrix}

\item{ncols}{Number of columns in the dose matrix}
}
\value{
Binary matrix where TRUE indicates adjacent doses
}
\description{
This function identifies dose combinations that are adjacent to the boundary
between acceptable and unacceptable doses.
}
\keyword{internal}
