% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual-summaryTables.R
\name{create_intervals}
\alias{create_intervals}
\title{Create intervals around target DLT probability}
\usage{
create_intervals(target_DLT_prob, interval_width)
}
\arguments{
\item{target_DLT_prob}{Target DLT probability}

\item{interval_width}{Width of the interval}
}
\value{
Vector of interval boundaries
}
\description{
Creates intervals around the target DLT probability for categorizing dose combinations.
}
