% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseConfigurator.R
\name{createDoseConfig}
\alias{createDoseConfig}
\title{Create a new DoseConfiguration object}
\usage{
createDoseConfig(
  drugCombiObject = NULL,
  isNonDecreasing = TRUE,
  currentConfig = NULL
)
}
\arguments{
\item{drugCombiObject}{An object of class DrugCombination}

\item{isNonDecreasing}{Logical indicating monotonicity direction (TRUE for non-decreasing)}

\item{currentConfig}{Optional numeric vector representing the initial configuration}
}
\value{
A DoseConfiguration object
}
\description{
This function creates a new DoseConfiguration object, which represents and manages
dose configurations in dose-finding studies.
}
\seealso{
\code{\link{DoseConfiguration-class}} for the class documentation
}
