% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugConstructor.R
\name{createDrug}
\alias{createDrug}
\title{Create a new Drug object}
\usage{
createDrug(drugData = NULL, name = NULL, doseCount = NULL)
}
\arguments{
\item{drugData}{Optional list containing drug information from a YAML file}

\item{name}{Optional character string for the drug name}

\item{doseCount}{Optional numeric value for the number of dose levels}
}
\value{
A Drug object or a list of Drug objects
}
\description{
This function creates one or more Drug objects based on the provided parameters.
}
