% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugCombiConstructor.R
\name{createDrugCombi}
\alias{createDrugCombi}
\title{Create a Drug Combination object}
\usage{
createDrugCombi(drugs = NULL, drug_A = NULL, drug_B = NULL)
}
\arguments{
\item{drugs}{List of \code{\link{Drug-class}} objects}

\item{drug_A}{Optional list with settings for first drug, containing 'name' and 'dose_levels'}

\item{drug_B}{Optional list with settings for second drug, containing 'name' and 'dose_levels'}
}
\value{
A new DrugCombi object
}
\description{
This function creates a new DrugCombi object for managing combinations of drugs
in dose-finding studies. It handles simple initialization with parameters or with
pre-existing Drug objects.
}
\details{
This function provides a convenient way to create a DrugCombi object either from
existing Drug objects or by specifying the parameters for two drugs directly.

If \code{drugs} is provided, it should be a list of Drug objects.
If \code{drug_A} and \code{drug_B} are provided, new Drug objects will be created with the specified parameters.
}
\seealso{
\code{\link{Drug-class}} for the drug class, \code{\link{createDrug}} for creating Drug objects
}
