% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialConstructor.R
\name{createPatientDataModel}
\alias{createPatientDataModel}
\title{Create a new PatientDataModel object}
\usage{
createPatientDataModel(
  drugDoseObject,
  admissibleRuleList = list(),
  selectionStrategyList = list(),
  startingDoseLevel = NULL,
  cohortSize = 3,
  maxCohorts = 20
)
}
\arguments{
\item{drugDoseObject}{A DrugCombination object}

\item{admissibleRuleList}{A list of admissibility rules}

\item{selectionStrategyList}{A list of selection strategies}

\item{startingDoseLevel}{The starting dose level (character)}

\item{cohortSize}{The number of patients per cohort}

\item{maxCohorts}{The maximum number of cohorts}
}
\value{
A PatientDataModel object
}
\description{
This function creates a new PatientDataModel object, which manages patient data and trial progression
in dose-finding studies.
}
