% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPEConstructor.R
\name{createPipeEstimator}
\alias{createPipeEstimator}
\title{Create a new PipeEstimator object}
\usage{
createPipeEstimator(
  validConfigs = NULL,
  epsilonTarget = 0.5,
  taper_type = NULL,
  custom_taper = NULL
)
}
\arguments{
\item{validConfigs}{List of valid dose configurations}

\item{epsilonTarget}{Target epsilon value after tapering}

\item{taper_type}{Type of tapering function ('linear', 'quadratic', or NULL)}

\item{custom_taper}{Custom tapering function (or NULL)}
}
\value{
A PipeEstimator object
}
\description{
This function creates a new PipeEstimator object, which implements the PIPE
methodology for dose-finding studies.
}
