% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugCombiConstructor.R
\docType{class}
\name{DrugCombination-class}
\alias{DrugCombination-class}
\alias{DrugCombi}
\title{Drug Combination Class}
\arguments{
\item{drugs}{List of Drug objects}

\item{drug}{A Drug object to add}

\item{combined}{Logical indicating whether to return the total number of combined dose levels (TRUE) or a vector of dose levels for each drug (FALSE)}

\item{doseCombi_selected}{Optional parameter to specify which dose combinations to return levels for}
}
\value{
List of dose combinations

Numeric value or vector of the number of dose levels

Vector of dose combination level identifiers
}
\description{
Initialize a new DrugCombination object

Add a new drug to the combination

Generate all possible combinations of drug dose levels

Update the combinations whenever there's a change in any of the Drug objects

Get all dose combinations

Get the number of dose levels

Get the level identifiers for dose combinations

Print a summary of the drug combination
}
\details{
A reference class that represents combinations of drugs.
This class manages objects of the \code{\link{Drug}} class to create
and manage combinations of different drugs at various dose levels.
}
\section{Fields}{

\describe{
\item{\code{drugs}}{List of Drug objects}

\item{\code{doseCombinations}}{List of generated dose combinations}
}}


\seealso{
\code{\link{Drug-class}} for the base drug class
}
