% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialConstructor.R
\name{find_closest_to_boundary}
\alias{find_closest_to_boundary}
\title{Find indices closest to a boundary}
\usage{
find_closest_to_boundary(vectorized_matrix, nrow, ncol)
}
\arguments{
\item{vectorized_matrix}{A vectorized matrix representation}

\item{nrow}{Number of rows in the matrix}

\item{ncol}{Number of columns in the matrix}
}
\value{
Vector of indices closest to the boundary
}
\description{
This helper function finds indices of 0s that are adjacent to 1s in a matrix.
}
\keyword{internal}
