% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies-admissible-BorderAdmissible.R
\name{findBorderIndices}
\alias{findBorderIndices}
\title{Find Border Indices in a Dose Configuration}
\usage{
findBorderIndices(A, matrixDims)
}
\arguments{
\item{A}{Binary vector representing the current dose configuration}

\item{matrixDims}{Vector containing the dimensions of the matrix}
}
\value{
Vector of indices representing border positions
}
\description{
This function identifies indices in a dose configuration that lie on the border
between acceptable and unacceptable doses based on neighbor sums.
}
\keyword{internal}
