% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies-admissible-Neighbour.R
\name{findNearestNeighbours}
\alias{findNearestNeighbours}
\title{Find Nearest Neighbours in a Matrix}
\usage{
findNearestNeighbours(pointIndex, matrixDims, degree = 1)
}
\arguments{
\item{pointIndex}{Integer index of the reference point in the matrix}

\item{matrixDims}{Vector containing the dimensions of the matrix (columns, rows)}

\item{degree}{Integer indicating the neighborhood degree (1 for immediate neighbors)}
}
\value{
Vector of indices representing neighbors of the reference point
}
\description{
This function identifies the indices of neighboring cells in a matrix given a point index.
}
\keyword{internal}
