% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual-plotFunctions.R
\name{plot_dlt_proportion}
\alias{plot_dlt_proportion}
\title{Plot DLT proportion}
\usage{
plot_dlt_proportion(sim_output, drugA, drugB, cohort_seq)
}
\arguments{
\item{sim_output}{Simulation output}

\item{drugA}{Drug A object}

\item{drugB}{Drug B object}

\item{cohort_seq}{Cohort sequence to visualize}
}
\value{
ggplot object
}
\description{
Creates a visualization of DLT proportions across dose combinations.
}
