% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual-plotFunctions.R
\name{prepare_step_data}
\alias{prepare_step_data}
\title{Prepare step data for visualization}
\usage{
prepare_step_data(sim_output, drugA, drugB, cohort_limit)
}
\arguments{
\item{sim_output}{Simulation output}

\item{drugA}{Drug A object}

\item{drugB}{Drug B object}

\item{cohort_limit}{Maximum cohort to include}
}
\value{
Data frame with step function coordinates
}
\description{
Prepares data for plotting step functions in the dose allocation visualization.
}
