% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseConfigurator.R
\name{slice_array_by_dimension}
\alias{slice_array_by_dimension}
\title{Slice an array by each dimension}
\usage{
slice_array_by_dimension(array)
}
\arguments{
\item{array}{The array to slice}
}
\value{
A list of lists, where each inner list contains slices along a dimension
}
\description{
This helper function slices a multi-dimensional array along each dimension.
}
\keyword{internal}
