% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies-selection-smallest_samplesize.R
\docType{class}
\name{SmallestSampleSizeStrategy-class}
\alias{SmallestSampleSizeStrategy-class}
\alias{SmallestSampleSizeStrategy}
\title{Smallest Sample Size Selection Strategy}
\arguments{
\item{admissibleDoses}{Vector of admissible doses}

\item{summaryStats}{Summary statistics for each dose combination}

\item{drugCombi}{DrugCombinationModel object containing dose information}
}
\value{
The selected dose combination
}
\description{
Select the next dose based on the smallest sample size
}
\details{
A reference class that implements a selection strategy for choosing the next dose
based on the smallest sample size. This strategy prioritizes dose levels that have
been tested on the fewest patients, promoting balanced exploration of the dose space.

The SmallestSampleSizeStrategy class implements the selectDose method by identifying
the admissible doses with the smallest posterior sample size (number of patients plus
prior pseudocounts). If multiple doses have the same minimum sample size, one is
randomly selected.
}

\examples{
\dontrun{
# Create base admissible rule class first if needed
library(gPIPE)

# Create a SmallestSampleSizeStrategy
smallest_sample_strategy <- SmallestSampleSizeStrategy$new()
}

}
\seealso{
\code{\link{SelectionStrategy}} for the base class
\code{\link{PosteriorProbabilityStrategy}}, \code{\link{InverseSqrtSampleSizeStrategy}},
\code{\link{InverseDistanceStrategy}}, \code{\link{EqualRandomisationStrategy}} for other selection strategies
}
