% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseModelConstructor.R
\name{createDrugCombinationModel}
\alias{createDrugCombinationModel}
\title{Create a new DrugCombinationModel object}
\usage{
createDrugCombinationModel(yamlConfig, family = "binomial")
}
\arguments{
\item{yamlConfig}{Path to YAML configuration file containing model parameters}

\item{family}{Character string specifying the distribution family ("binomial" or "gaussian")}
}
\value{
A DrugCombinationModel object
}
\description{
This function creates a new DrugCombinationModel object, which implements
the statistical model for a trial.
}
